requirements.txt
# Web API
fastapi==0.110.1
uvicorn[standard]==0.29.0
httpx==0.27.0

# NLP & Transformers
transformers==4.40.0
torch==2.2.2
tokenizers==0.19.1

# NLP complémentaire
spacy==3.7.2
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl

# Schémas & validation
pydantic==2.7.1
pydantic-core==2.18.2

# Utilitaires divers
python-dotenv==1.0.1

loguru==0.7.0
rich==13.7.1